refactor: space 생성, 수정 비지니스 로직을 커스텀훅으로 분리한다
